feat: implement multi-role support and role array migration
- Migrate database role column from VARCHAR to TEXT[] array
- Support users with multiple roles (e.g., student + tutor simultaneously)
- Add role switcher UI for multi-role users in Layout component
- Update all backend endpoints to check role arrays properly
- Update all frontend components to use role.includes() instead of role ===
- Fix permission checks across admin, coordinator, schedule_preferences, sessions, materials, forum, scheduling APIs
- Add any_() operator for PostgreSQL array queries
- Add TutorStatistics page showing existing courses count per subject
- Fix Admin page to display role arrays properly
- Update authStore to support currentMode and role switching
- Maintain backward compatibility with single role strings

+++ backend/app/schemas/tutor.py
@@ -36,7 +36,7 @@ class TutorResponse(TutorBase):
     """Tutor response DTO"""
     tutor_id: int
     user_id: int
-    rating: Decimal
+    rating: float
     total_sessions: int
     is_verified: bool
     verified_at: Optional[datetime]
